drm: Introduce a drm_blob abstraction
This becomes the base class of the existing wlserver blob structs, and
will be used for more DRM state tracking in future.

The copy and move constructors are deleted to ensure we don't mishandle
lifetimes.

One potential evolution direction is to make this take the blob data
struct as a generic parameter and make it handle creation as well, but
keep it simple for now.